AsynchronousEventManager Complete Reimplementation
Completely reimplemented AsynchronousEventManager to implement Runnable rather than Tickable, thus removing dependency on MtgSaberLib_Threads. The new implementation is based on the classic Readers/Writers problem, except that here there are many event producers pushing on to one event consumer*. This consumer then distributes the events to its internal table of subscribers. It is meant to run on its own thread, and will not function unless it is offered the thread that will run it. It will block execution until there is an event to distribute, and if there are multiple producers waiting to push events, it will wait for all of them to finish before distributing the events. This puts the priority on the production of events, rather than the latency of their distribution.